refactor(/asis): controller 인터페이스 기반 레거시 @Deprecated 처
