+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/asis/Controller.java
@@ -3,6 +3,7 @@ package com.interface21.webmvc.servlet.mvc.asis;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
+@Deprecated
 public interface Controller {
     String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception;
 }
