+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/ControllerHandlerAdapterTest.java
@@ -20,6 +20,7 @@ import org.mockito.ArgumentCaptor;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
+import com.interface21.webmvc.servlet.mvc.asis.ControllerHandlerAdapter;
 import com.interface21.webmvc.servlet.mvc.tobe.HandlerExecution;
 
 import samples.TestAnnotationController;
